requirements.txt
requests
python-dotenv
fastapi
uvicorn
pydantic
openapi-spec-validator
faiss-cpu
sentence-transformers